Add code to project based on future expected supply
But then disable it, it's not particularly enlightening (yet) because AZ
second doses aren't kicking in and future expected supply is basically
no change from the present.

Since things are in a state of flux, will leave this disabled for now
and re-enable if there are announced changes to supply.