Add bipartite model of NSW epidemic
Treating the LGAs of concern and the rest of NSW separately.

The total population of the two regions is not accurate, as Sydney's
population is used for both of them. That's fine for now since we're
not getting close to significant population immunity. If things get
worse we'll need to put in the correct populations.