Remove unused code from nsw.py
Prior to copying stuff over to vic-2021.py

Not gonna actually factor things into functions or anything like that,
am still going to copy-paste, but this will make it easier to keep
changes in sync.

Usually bad practice to copy-paste large chunks of code like this, but
there are so many fiddly small differences that encapsulating into
re-usable functions is hard too.