Use expected dk_dt for fits used for padding.
Incorporate expected effect of vaccination into this fit so that we can
extract a growth rate more likely to match today's growth rate rather
than the average over the last 2 weeks.

Also, temporarily reduce number of points used in fit to 10, to reduce
influence of data from before the grand-final-weekend surge. Will
increase it again once that data is out of the window.