NZ: use press release number for the latest day
official dataset seems to be 2 days out of date now, so we need to do
something like this to not be biased downward in our R_eff estimates.

Other alternative is to discard an extra day and be a day out of date,
but that is not very appealing, and I don't think using a single
9am-9am datapoint that doesn't overlap with our other reporting periods
will introduce a bias - it is still a representative 24-hour period.