Ensure we get an up to date NZ MoH page
Today they were having what looked like caching issues with load
balancing, and half the page gets returned an old page.

Also, use the table at the bottom of the page for net change in cases,
instead of net change in active. The latter was wrong - I think it was
only temporarily right because recoveries were not being reported for a
day or two.