Add "R_eff in Greater Sydney vs rest of NSW" section
I haven't included the output files in this commit as my computer seems to do something different with the float precisions compared to the images you've generated. Let me know if you want me include the PNG/SVG files and changes to latest_nsw_stats.json in the commit as well.