Resolve inconsistency between delta and exponential generation distributions
Was previously treating these as the same, which was fine for small
R_eff.

We were determining R_eff using an implicit delta generation
distribution, but projecting forward using an exponential distribution.

Now that R_eff is big, the difference between the two distributions is
large and we need to do the conversion to not project unrealistically
large cases.